Fixed signedness, using cgalwraps in Armadillo.cpp

+++ src/mesh/pe/rigid_body/ConvexPolyhedronFactory.cpp
@@ -146,7 +146,7 @@ TriangleMeshUnion* createApproximateNonConvexUnion( BodyStorage& globalStorage, 
    std::vector<TriangleMesh> convexParts = ConvexDecomposer::approximateConvexDecompose(mesh, max_concavity);
 
    // Centrate parts an add them to the union
-   for(int part = 0; part < (int)convexParts.size(); part++){
+   for(size_t part = 0; part < convexParts.size(); part++){
       Vec3 centroid = mesh::toWalberla( mesh::computeCentroid( convexParts[part] ) );
       mesh::translate( convexParts[part], -centroid );
       createConvexPolyhedron(un, uid, centroid+gpos, convexParts[part], material,
